+++ CHANGES.md
@@ -5,3 +5,15 @@ All notable changes to this project will be documented in this file.
 ## [0.1.0] - 2020-12-14
 
 - Initial commit
+
+## [0.1.1] - 2021-01-08
+
+- Added "curly": true rule
+
+## [0.1.2] - 2021-02-08
+
+- "explicit-function-return-type" TS override
+
+## [0.1.3] - 2021-03-04
+
+- Add indentation for SwitchCase